amend-ocr-1/requirements.txt
boto3==1.38.33
openai==1.88.0
pandas==2.3.0
pillow
PyMuPDF